style line chart
