+++ frontend/src/components/analytics/charts/index.ts
@@ -1,0 +1,3 @@
+import LineChart from './LineChartComponent';
+
+export { LineChart };
